Fixed issues with "Developer - release" configuration; implementation of AvHTunnelToServer.
git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@26 67975925-1194-0748-b3d5-c16f83f1a3a1

+++ main/source/mod/AvHNexusServer.h
@@ -8,8 +8,8 @@ typedef struct entvars_s entvars_t;
 
 namespace AvHNexus
 {
-	bool send(entvars_t* const pev, const unsigned char* data, const unsigned int length);
-	bool recv(entvars_t* const pev, const char* data, const unsigned int length);
+	bool send(entvars_t* const pev, const unsigned char* data, const size_t length);
+	bool recv(entvars_t* const pev, const char* data, const size_t length);
 
 	void handleUnauthorizedJoinTeamAttempt(const edict_t* edict, const unsigned char team_index);
 	string getNetworkID(const edict_t* edict);
